+++ process_profile_update.php
@@ -589,7 +589,7 @@ if ($action == 'remove_photo') {
         $stmt->execute([$current_user_id]);
         $old_image = $stmt->fetchColumn();
         
-        if ($old_image && file_exists($old_image)) {
+        if ($old_image && !preg_match('#^https?://#', $old_image) && file_exists($old_image)) {
             unlink($old_image);
         }
